Fix SSE stream interval config
- Changed sleep function to usleep for more precise control over the interval in StreamController.
- Updated configuration to reflect that the hot reload interval is now in microseconds instead of seconds, ensuring clarity in the expected value type.

+++ config/horizonhub.php
@@ -114,7 +114,7 @@ return [
     | This is not a full browser reload interval.
     |
     */
-    'hot_reload_interval' => (int) env('HORIZON_HUB_HOT_RELOAD_INTERVAL', 0.5),
+    'hot_reload_interval' => (float) env('HORIZON_HUB_HOT_RELOAD_INTERVAL', 0.5),
 
     /*
     |--------------------------------------------------------------------------
